requirements.txt
httplib2==0.9.2
pkg-resources==0.0.0
python-dwca-reader==0.9.2
-e git+git@github.com:stan3/pywikibot-core.git@beta_commons#egg=pywikibot
xlrd==1.0.0
